% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sparkler.R
\name{sparkTex}
\alias{sparkTex}
\title{Convert ggplot object to LaTex}
\usage{
sparkTex(gg_obj, width = 4)
}
\arguments{
\item{gg_obj}{Ggplot object.}

\item{width}{Width, in em, for your sparkline.}
}
\description{
This function converts a ggplot object to latex code. Requires \usepackage{sparklines} in your latex header.
}
\examples{
sparkTex()
}
\keyword{sparkline}

