% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireworks.R
\name{fireworks}
\alias{fireworks}
\title{Launch Fireworks}
\usage{
fireworks(
  duration = 5,
  speed = 1,
  width = "100\%",
  height = "0px",
  elementId = NULL
)
}
\arguments{
\item{duration}{Number of seconds to run. Default is 5.}

\item{speed}{Number. Speed of simulation (default 1).}

\item{width, height}{Must be valid CSS unit. Leave as default for overlay.}

\item{elementId}{Optional ID.}
}
\description{
Triggers a fireworks display on screen.
}
