% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{Start Weather Effect}
\usage{
weather(
  type = "snow",
  density = 1,
  speed = 1,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{type}{Character. One of "snow", "rain", "meteor", or "none".}

\item{density}{Numeric. Multiplier for particle count (default 1).}

\item{speed}{Numeric. Multiplier for movement speed (default 1).}

\item{elementId}{Optional ID.}
}
\description{
Start Weather Effect
}
